refactor: shared 및 app 레이어 사용으로 import 경로 업데이트
변경 사항:
- index.tsx에서 app/providers와 app/router 사용
- 모든 페이지에서 shared 레이어로 import 변경
  * pages/Main, CreateRoom, GameRoom
- Game 컴포넌트에서 shared 레이어 사용
- components에서 shared 레이어 사용
- socket/index.ts에서 shared/config 사용
- shared/index.ts에서 모든 shared 모듈 export

이유:
- FSD 아키텍처에 맞게 모든 import 경로를 shared/app 레이어로 통일
- 코드 일관성 및 유지보수성 향상
