+++ client/src/shared/index.ts
@@ -1,3 +1,5 @@
 export * from './ui';
 export * from './lib';
+export * from './config';
+export * from './api';
