+++ client/src/socket/index.ts
@@ -5,4 +5,4 @@ export { gameSocket } from './gameSocket';
 export type { MessageType, MessageCallback, SocketMessage } from './types';
 
 // 상수
-export { SOCKET_SERVER_URL, SOCKET_ENDPOINTS, SOCKET_TOPICS } from './constants';
+export { SOCKET_SERVER_URL, SOCKET_ENDPOINTS, SOCKET_TOPICS } from '../shared/config';
